- Rename groupId to au.edu.federation.caliko.demo to be consistent with Maven module names
