+++ pom.xml
@@ -44,7 +44,7 @@
         <version>${project.version}</version>
       </dependency>
       <dependency>
-        <groupId>au.edu.federation.caliko.alansley</groupId>
+        <groupId>au.edu.federation.caliko.demo</groupId>
         <artifactId>caliko-demo</artifactId>
         <version>${project.version}</version>
       </dependency>
